# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API E-commerce Segura con Swagger Completo",
        "description": "API RESTful para la gestión de un e-commerce: registro y autenticación de clientes mediante tokens, catálogo de productos, creación de órdenes de compra y procesamiento de pagos con Stripe.",
        "contact": {
            "email": "soporte@ecommerce-api.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listado público del catálogo de productos",
                "description": "Devuelve los productos activos, con paginación y búsqueda opcional por nombre o descripción. No requiere autenticación.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Texto a buscar en nombre o descripción",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de resultados por página (por defecto 15)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de productos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un producto",
                "description": "Crea un nuevo producto en el catálogo. Requiere un token de un usuario administrador.",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop HP Pavilion 15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Laptop de 15 pulgadas, 16GB RAM, 512GB SSD"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 699.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://ejemplo.com/laptop.jpg"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (se requiere rol de administrador)"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Detalle de un producto",
                "description": "Devuelve la información completa de un producto por su identificador. No requiere autenticación.",
                "operationId": "db00b523e69d5dbc09c6edad76374bdc",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del producto"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "description": "Actualiza parcial o totalmente un producto existente. Requiere un token de un usuario administrador.",
                "operationId": "f47cf55f4f3e0b134b5e7d9a8373725d",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop HP Pavilion 15 (2026)"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Actualización de memoria a 32GB RAM"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 749.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 18
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://ejemplo.com/laptop.jpg"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (se requiere rol de administrador)"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "description": "Elimina un producto del catálogo. Requiere un token de un usuario administrador.",
                "operationId": "ac878db81fdb1333b21fc79e574a5a17",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (se requiere rol de administrador)"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stripe/webhook": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Webhook de Stripe (uso exclusivo de Stripe)",
                "description": "Endpoint invocado directamente por Stripe para notificar eventos de pago (payment_intent.succeeded, payment_intent.payment_failed). La firma de la petición se verifica con STRIPE_WEBHOOK_SECRET. No requiere autenticación de usuario, pero sí una firma válida de Stripe. No debe invocarse manualmente.",
                "operationId": "2ca16d2f99bc06cb8a0075bb4dbf33a6",
                "responses": {
                    "200": {
                        "description": "Evento procesado correctamente"
                    },
                    "400": {
                        "description": "Firma de webhook inválida o payload malformado"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtenido en /api/register o /api/login. Enviar como: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum Personal Access Token",
                "scheme": "bearer"
            }
        }
    },
...
}